requirements.txt
amqp==5.2.0
asgiref==3.8.1
billiard==4.2.1
branca==0.8.0
Brotli==1.1.0
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
channels==4.1.0
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
crispy-bootstrap5==2024.10
crispy-tailwind==1.0.3
cssselect2==0.7.0
Django==5.1.2
django-admin-generator==2.6.0
django-allauth==65.0.2
django-appconf==1.0.6
django-countries==7.6.1
django-crispy-forms==2.3
django-extensions==3.2.3
django-filter==24.3
django-haystack==3.3.0
django-imagekit==5.0.0
django-phonenumber-field==8.0.0
django-ratelimit==4.1.0
django-tables2==2.7.0
django-tinymce==4.1.0
django-utils2==3.0.2
django-widget-tweaks==1.5.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
folium==0.17.0
fonttools==4.54.1
html5lib==1.1
idna==3.10
Jinja2==3.1.4
kombu==5.4.2
MarkupSafe==3.0.2
numpy==2.1.2
packaging==24.1
phonenumbers==8.13.47
pilkit==3.0
pillow==11.0.0
prompt_toolkit==3.0.48
psycopg2-binary==2.9.10
pycparser==2.22
pydyf==0.11.0
PyJWT==2.9.0
pyphen==0.16.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-utils==3.9.0
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
tinycss2==1.4.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
weasyprint==62.3
webencodings==0.5.1
Whoosh==2.7.4
xyzservices==2024.9.0
zopfli==0.2.3.post1
